ARM: Stricten arguments to virt_to_pfn()
The kernel convention is that the argument passed to virt_to_pfn()
(virtual address to page frame number) needs to be an unsigned long,
but ARM has avoided this by including a cast to (unsigned long)
inside the definition of virt_to_pfn() itself. This creates problems
if we want to make virt_to_pfn() an actual function with a typed argument,
so fix up the offending call sites.

+++ arch/arm/include/asm/dma-direct.h
@@ -28,7 +28,7 @@ static inline unsigned long dma_to_pfn(struct device *dev, dma_addr_t addr)
 static inline dma_addr_t virt_to_dma(struct device *dev, void *addr)
 {
 	if (dev)
-		return pfn_to_dma(dev, virt_to_pfn(addr));
+		return pfn_to_dma(dev, virt_to_pfn((unsigned long)addr));
 
 	return (dma_addr_t)__virt_to_bus((unsigned long)(addr));
 }
